added warnings for when a brush doesn't have a terrain reference
There are two warnings.

the first one appears below the terrain field where the reference should go, explaining the field must have something in it or else the brush will not modify any terrain.

the second warning appears in the console when the scene with the brush enters play mode, stating the above while also giving the name of the object causing the warning.

Time Taken: 40 minutes